move generators into utils
