+++ src/api/composeDidWebs.ts
@@ -1,5 +1,5 @@
 import { Did } from '../core/Did.js';
-import { generateDid } from '../core/generators/generateDid.js';
+import { generateDid } from '../utils/generators/generateDid.js';
 
 export const composeDidWebs = (
   host: string,
